chore: app, componets 디렉토리 src 내로 이동
