--- apps/admin/postcss.config.js
@@ -1,3 +1,0 @@
-import { postcssConfig } from "@repo/tailwind-config/postcss";
-
-export default postcssConfig;
