+++ apps/admin/src/app/globals.css
@@ -1,4 +1,4 @@
-@import "@repo/ui/styles.css";
+@import "@workspace/ui/styles.css";
 
 :root {
   --background: #ffffff;
