+++ apps/admin/src/app/page.tsx
@@ -1,0 +1,17 @@
+import { Button } from "@workspace/ui/components/button";
+import { Progress } from "@workspace/ui/components/progress";
+
+export default function Page() {
+  return <div>
+    <Button variant="default">다음</Button>
+    <Button variant="disabled">다음</Button>
+    <Button variant="outline">다음</Button>
+    <Button variant="onb_select">다음</Button>
+    <Button variant="onb_selected">다음</Button>
+    <Button variant="unselected">다음</Button>
+    <Button variant="period_selected">다음</Button>
+    <Button variant="kakao">다음</Button>
+    <Button variant="black">다음</Button>
+    <div></div>
+  </div>;
+}
