--- apps/user/app/page.tsx
@@ -1,6 +1,0 @@
-import { Button } from "@workspace/ui/components/button";
-import { Badge } from "@workspace/ui/components/badge";
-
-export default function Page() {
-  return <div></div>;
-}
