+++ apps/user/src/app/layout.tsx
@@ -1,9 +1,28 @@
-import "./globals.css";
+import { Geist, Geist_Mono } from "next/font/google";
 
-export default function RootLayout({ children }: { children: React.ReactNode }) {
+import "@workspace/ui/globals.css";
+import { Providers } from "@/components/providers";
+
+const fontSans = Geist({
+  subsets: ["latin"],
+  variable: "--font-sans",
+});
+
+const fontMono = Geist_Mono({
+  subsets: ["latin"],
+  variable: "--font-mono",
+});
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
   return (
-    <html lang="en">
-      <body>{children}</body>
+    <html lang="en" suppressHydrationWarning>
+      <body className={`${fontSans.variable} ${fontMono.variable} font-sans antialiased`}>
+        <Providers>{children}</Providers>
+      </body>
     </html>
   );
 }
