+++ apps/user/src/app/page.tsx
@@ -1,3 +1,7 @@
+import { Button } from "@workspace/ui/components/button";
+
 export default function Page() {
-  return <div></div>;
+  return <div>
+        <Button></Button>
+  </div>;
 }
