+++ packages/ui/eslint.config.mjs
@@ -1,4 +1,4 @@
-import { config } from "@repo/eslint-config/react-internal";
+import { config } from "@workspace/eslint-config/react-internal";
 
 /** @type {import("eslint").Linter.Config} */
 export default config;
